カメラ追従改善、エイムアシスト追加、コード整形
カメラ追従ロジックを改善し、エイムアシスト機能を追加。
また、コードフォーマットの統一やリファクタリングを実施。

- カメラ追従ロジックの補間アルゴリズムを調整し、動作をスムーズ化。
- エイムアシスト関連メソッドとパラメータを追加・調整。
- ターゲット位置取得のキャッシュ機構を導入し、パフォーマンス向上。
- デバッグ用カメラ位置ログ出力を追加（DEBUGビルド限定）。
- カメラ初期位置設定のバグを修正。
- `PlayerMachineGun::Shoot`の弾ばらつき計算ロジックを改善。
- マシンガンの弾速やばらつきの定数を調整。
- `CollisionManager`の衝突描画機能を追加。
- `BaseEnemy`や`GameScene`など複数ファイルでコード整形を実施。
- 不要なコードやコメントを削除し、可読性を向上。
- `imgui.ini`でUIレイアウトを調整。

+++ application/scene/GameScene.cpp
@@ -33,7 +33,7 @@ void ShowGradientEditor() {
 
 	// 制御点の編集
 	for (size_t i = 0; i < gradientPoints.size(); ++i) {
-		ImGui::PushID(i);
+		ImGui::PushID(static_cast<int>(i));
 		ImGui::DragFloat("Position", &gradientPoints[i].position, 0.01f, 0.0f, 1.0f);
 		ImGui::ColorEdit4("Color", (float *)&gradientPoints[i].color);
 		ImGui::PopID();
@@ -634,6 +634,7 @@ void GameScene::Update() {
 	ShowGradientEditor();
 
 	hud_->DrawImGui();
+
 #endif
 
 	lineDrawer_->Update();
